[TASK] Move downloads to dedicated controller
Also the tests are enhanced to avoid download issues.
